[API] Finish rename of the Presentation project (on a folder root level).
